Add logic to trim alleles for uncomplicated cases.
In selected cases where the input gvcf record has a more padded ref (and alts) as compared to the unified site, we can trim off the extra padding bases at the end such that some of the input alleles will be translatable to the leaner unified representation. Note that all ALT calls in sites where the input record has range differing from the unified will be lost previously. This commit should reduce loss in some of these cases